@@ -1,4 +1,4 @@
-class CreateSubjectAlternateNames < ActiveRecord::Migration
+class CreateSubjectAlternateNames < ActiveRecord::Migration[4.2]
   def change
     create_table :subject_alternate_names do |t|
       t.string :name, null: false
