@@ -1,4 +1,4 @@
-class CreateRevocationEndpoints < ActiveRecord::Migration
+class CreateRevocationEndpoints < ActiveRecord::Migration[4.2]
   def change
     create_table :revocation_endpoints do |t|
       t.integer :public_key_id, null: false
