@@ -1,4 +1,4 @@
-class CreateServices < ActiveRecord::Migration
+class CreateServices < ActiveRecord::Migration[4.2]
   def change
     create_table :services do |t|
       t.string :type, null: false
