@@ -1,4 +1,4 @@
-class CreateKeyUsages < ActiveRecord::Migration
+class CreateKeyUsages < ActiveRecord::Migration[4.2]
   def change
     create_table :key_usages do |t|
       t.integer :public_key_id, null: false
