@@ -1,4 +1,4 @@
-class CreateOAuthProviders < ActiveRecord::Migration
+class CreateOAuthProviders < ActiveRecord::Migration[4.2]
   def change
     create_table :o_auth_providers do |t|
       t.string :name, null: false
