@@ -1,4 +1,4 @@
-class CreateAgentsServices < ActiveRecord::Migration
+class CreateAgentsServices < ActiveRecord::Migration[4.2]
   def change
     create_table :agents_services do |t|
       t.integer :agent_id, null: false
