@@ -1,4 +1,4 @@
-class CreateAgentTags < ActiveRecord::Migration
+class CreateAgentTags < ActiveRecord::Migration[4.2]
   def change
     create_table :agent_tags do |t|
       t.integer :agent_id, null: false
