@@ -1,4 +1,4 @@
-class CreateCsrSan < ActiveRecord::Migration
+class CreateCsrSan < ActiveRecord::Migration[4.2]
   def change
     create_table :csr_sans do |t|
       t.integer :certificate_sign_request_id, null: false
